requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
numpy
scikit-learn
joblib